Move message list and message compose tests
